requirements.txt
streamlit
opencv-contrib-python
Pillow
numpy
